Extract DecoratingRobot from TracingRobot
